Move Matchers to CatchUtils folder and fix inclusion order
