+++ .clang-format
@@ -99,10 +99,12 @@ IncludeCategories:
     Priority:        5
   - Regex:           '^"vecmath/.*'
     Priority:        5
+  - Regex:           '^"CatchUtils/.*'
+    Priority:        10
   - Regex:           '^<catch2/.*'
-    Priority:        10
+    Priority:        11
   - Regex:           '^"Catch2.h"'
-    Priority:        10
+    Priority:        11
   - Regex:           '^".*'
     Priority:        3
   - Regex:           '^<.+/.*'
