+++ common/test/CMakeLists.txt
@@ -4,14 +4,14 @@ set(COMMON_TEST_UTILS_SOURCE
         "${COMMON_TEST_SOURCE_DIR}/Assets/EntityDefinitionTestUtils.cpp"
         "${COMMON_TEST_SOURCE_DIR}/Assets/EntityDefinitionTestUtils.h"
         "${COMMON_TEST_SOURCE_DIR}/Catch2.h"
+        "${COMMON_TEST_SOURCE_DIR}/CatchUtils/Matchers.cpp"
+        "${COMMON_TEST_SOURCE_DIR}/CatchUtils/Matchers.h"
         "${COMMON_TEST_SOURCE_DIR}/IO/TestEnvironment.cpp"
         "${COMMON_TEST_SOURCE_DIR}/IO/TestEnvironment.h"
         "${COMMON_TEST_SOURCE_DIR}/IO/TestFileSystem.cpp"
         "${COMMON_TEST_SOURCE_DIR}/IO/TestFileSystem.h"
         "${COMMON_TEST_SOURCE_DIR}/IO/TestParserStatus.cpp"
         "${COMMON_TEST_SOURCE_DIR}/IO/TestParserStatus.h"
-        "${COMMON_TEST_SOURCE_DIR}/Matchers.cpp"
-        "${COMMON_TEST_SOURCE_DIR}/Matchers.h"
         "${COMMON_TEST_SOURCE_DIR}/QtPrettyPrinters.h"
         "${COMMON_TEST_SOURCE_DIR}/RunAllTests.cpp"
         "${COMMON_TEST_SOURCE_DIR}/TestLogger.cpp"
@@ -27,6 +27,7 @@ set(COMMON_TEST_SOURCE
         "${COMMON_TEST_SOURCE_DIR}/Assets/tst_EntityModel.cpp"
         "${COMMON_TEST_SOURCE_DIR}/Assets/tst_ModelDefinition.cpp"
         "${COMMON_TEST_SOURCE_DIR}/Assets/tst_Palette.cpp"
+        "${COMMON_TEST_SOURCE_DIR}/CatchUtils/tst_Matchers.cpp"
         "${COMMON_TEST_SOURCE_DIR}/EL/tst_EL.cpp"
         "${COMMON_TEST_SOURCE_DIR}/EL/tst_Expression.cpp"
         "${COMMON_TEST_SOURCE_DIR}/EL/tst_Interpolator.cpp"
