+++ common/test/src/IO/tst_DiskFileSystem.cpp
@@ -29,12 +29,13 @@
 #include "IO/TestEnvironment.h"
 #include "IO/TraversalMode.h"
 #include "Macros.h"
-#include "Matchers.h"
 
 #include "kdl/regex_utils.h"
 
 #include <algorithm>
 #include <filesystem>
+
+#include "CatchUtils/Matchers.h"
 
 #include "Catch2.h"
