+++ common/test/src/IO/tst_DiskIO.cpp
@@ -29,13 +29,14 @@
 #include "IO/TestEnvironment.h"
 #include "IO/TraversalMode.h"
 #include "Macros.h"
-#include "Matchers.h"
 
-#include "kdl/regex_utils.h"
+#include <kdl/regex_utils.h>
 
 #include <algorithm>
 #include <filesystem>
 #include <fstream>
+
+#include "CatchUtils/Matchers.h"
 
 #include "Catch2.h"
