+++ common/test/src/Model/tst_BrushRegression.cpp
@@ -24,7 +24,6 @@
 #include "IO/DiskIO.h"
 #include "IO/NodeReader.h"
 #include "IO/TestParserStatus.h"
-#include "Matchers.h"
 #include "Model/Brush.h"
 #include "Model/BrushBuilder.h"
 #include "Model/BrushFace.h"
@@ -50,6 +49,8 @@
 #include <fstream>
 #include <string>
 #include <vector>
+
+#include "CatchUtils/Matchers.h"
 
 #include "Catch2.h"
